Cleaned up interface definitions. Moved from global .d.ts files into local files. Export and import where necessary. Will use this pattern going forward.
