+++ src/components/Amendments/Amendments.tsx
@@ -1,4 +1,14 @@
 import * as React from "react";
+import { FoundationNode } from '../Foundation';
+
+interface AmendmentsProps {
+  amendmentsNodes: Array<FoundationNode>,
+  textIsHighlighted: boolean,
+  onBackClick: () => void,
+  onClearClick: () => void,
+  onSetClick: () => void,
+  onMouseUp: () => void
+}
 
 class Amendments extends React.Component<AmendmentsProps, void> {
   constructor(props: AmendmentsProps){
@@ -23,7 +33,7 @@ class Amendments extends React.Component<AmendmentsProps, void> {
             <p>&lt;This is just a placeholder block for section links to autoscroll to particular parts of the amendments&gt;</p>
           </div>
           <div className="amendments__text" onMouseUp={this.props.onMouseUp}>
-            {this.props.amendmentsNodes.map(function(element: MyReactComponentObject, index: number){
+            {this.props.amendmentsNodes.map(function(element: FoundationNode, index: number){
               element.props['key'] = index.toString();
               return(
                 React.createElement(element.component, element.props, element.innerHTML)
