+++ src/components/Foundation/index.ts
@@ -1,3 +1,4 @@
-import Foundation from './Foundation';
+import Foundation, { FoundationNode, FoundationNodeProps, FoundationTextTypes } from './Foundation';
 
+export { FoundationNode, FoundationNodeProps, FoundationTextTypes };
 export default Foundation;
